Create garbage-value.cpp
An array garbage value in C++ is a value that is stored in an array element but is not initialized properly. This can happen when an array element is declared but not assigned a value, or when an array element is assigned a value that is outside of its range.